requirements.txt
git+https://github.com/Theano/Theano.git#egg=Theano
python-mnist
numpy